# docs/swagger.yaml
basePath: /api
definitions:
  models.CheckoutItem:
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
    type: object
  models.CheckoutRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/models.CheckoutItem'
        type: array
    type: object
  models.Product:
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
      stock:
        type: integer
    type: object
  models.Transaction:
    properties:
      created_at:
        type: string
      details:
        items:
          $ref: '#/definitions/models.TransactionDetail'
        type: array
      id:
        type: integer
      total_amount:
        type: integer
    type: object
  models.TransactionDetail:
    properties:
      id:
        type: integer
      product_id:
        type: integer
      product_name:
        description: omitempty = tidak muncul di JSON kalau kosong
        type: string
      quantity:
        type: integer
      subtotal:
        type: integer
      transaction_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: API untuk sistem manajemen kasir/penjualan
  title: Kasir API Documentation
  version: "1.0"
paths:
  /checkout:
    post:
      consumes:
      - application/json
      description: Create a new transaction from checkout items (requires API key)
      parameters:
      - description: Checkout items
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CheckoutRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Transaction created
          schema:
            $ref: '#/definitions/models.Transaction'
        "400":
          description: Bad request or empty items
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Process checkout
      tags:
      - Transactions
  /produk:
    get:
      description: Retrieve all products with optional name filter
      parameters:
      - description: Product name filter
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of products
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get all products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Add a new product to inventory
      parameters:
      - description: Product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Product created
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad request
          schema:
            type: string
      summary: Create a new product
      tags:
      - Products
  /produk/{id}:
    delete:
      description: Remove a product from inventory (requires API key)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Deletion success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete a product
      tags:
      - Products
    get:
      description: Retrieve a specific product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product found
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Product not found
          schema:
            type: string
      summary: Get product by ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Update product information (requires API key)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "200":
          description: Product updated
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update a product
      tags:
      - Products
  /report/hari-ini:
    get:
      description: Get sales report for today or specific date range
      parameters:
      - description: 'Start date (format: 2026-01-01)'
        in: query
        name: start_date
        type: string
      - description: 'End date (format: 2026-01-01)'
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sales report summary
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request or invalid date format
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get sales report
      tags:
      - Reports
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: API Key untuk autentikasi
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"